fix(usage): include failed-call cost; refactor usage repo; tidy docs
Address PR #121 review feedback.

* Cost — drop status='ok' / not-all-failed filters on total_cost_usd in
  both per-LLM-call and per-invocation queries so the total reflects
  what was actually spent (including failed attempts the provider
  billed). Distributions and token totals stay ok-only so failures
  cannot skew quantiles. Numerically a no-op on today's data (the
  tracking adapter writes cost=0 on error) but the SQL is now robust
  if a future adapter records real cost on an error.
* Refactor — extract SqlAlchemyUsageRepository and its SQL-builder /
  row-parsing helpers from qfa.adapters.db into a new
  qfa.adapters.usage_repository module. Helpers become @staticmethod
  on the class. qfa.adapters.db keeps only the connection/session
  machinery, AAD token provider, and the llm_calls table declaration.
* Docs —
  * docs/rest-api/index.md: rewrite for API consumers; drop
    "inherited fields" / "per-call_id aggregation" jargon and the
    "unchanged vs. pre-#91" framing that assumed prior context.
  * _grouping_sets_clause: brief explanation of GROUPING SETS.
  * _select_per_llm_call: describe what the function emits now,
    not "extension of the previous _select_totals".
  * UsageMetrics / OperationStats / UsageStats: drop NumPy-style
    Attributes blocks that duplicated Field(description=...); the
    Field descriptions are the single source of truth and surface
    in OpenAPI.
  * routes_usage.py: update cost / failed_calls paragraph.
* Tests — integration test for the cost policy seeds an error row
  with non-zero cost so the new behaviour is locked in.
