+++ docs/rest-api/index.md
@@ -23,18 +23,18 @@ All endpoints except `GET /v1/health` require `Authorization: Bearer <key>`.
 | `GET` | `/v1/usage/all` | Cross-tenant stats (requires `is_superuser=true`) |
 | `GET` | `/v1/health` | Liveness probe; no auth |
 
-## Usage endpoint response shape (issue #91)
+## Usage endpoint response shape
 
 Both `GET /v1/usage` and `GET /v1/usage/all` return aggregated stats in two parallel views:
 
-- **Per-invocation** (inherited top-level fields): one count per distinct API call (per-`call_id` aggregation). Multi-LLM-call operations such as `/v1/assign_codes` collapse to a single invocation.
-- **Per-LLM-call** (`llm_call_stats`): one count per LLM call attempt — identical to the pre-#91 behaviour.
+- **Per REST API call** (top-level fields): each distinct call to one of the analysis endpoints (`/v1/analyze`, `/v1/summarize`, `/v1/summarize-aggregate`, `/v1/assign_codes`) counts as one. An endpoint like `/v1/assign_codes` that fans out to several LLM calls internally still shows up as a single entry here.
+- **Per LLM call** (`llm_call_stats`): each individual LLM provider call counts as one. Use this view when you want to see raw provider traffic — for example to compute the LLM-calls-per-API-call ratio (`llm_call_stats.total_calls / total_calls`).
 
-`operations` carries a per-operation breakdown of the same data, sorted by `total_cost_usd` desc (ties: operation asc), with empty operations omitted. Every per-operation block also carries its own `llm_call_stats` so clients can compute fan-out (`llm_call_stats.total_calls / total_calls`) per operation.
+`operations` carries a per-operation breakdown of the same data, sorted by `total_cost_usd` desc (ties: operation asc), with operations that had no traffic in the window omitted. Every per-operation entry also carries its own `llm_call_stats`.
 
-`total_cost_usd`, `input_tokens.total`, and `output_tokens.total` are numerically unchanged. `total_calls`, `failed_calls`, and every distribution field changed semantics for multi-LLM-call operations — read `llm_call_stats` for the previous semantics.
+`total_cost_usd` sums every row in the window — including failed attempts that incurred a real cost — so the figure reflects what was actually spent. Distributions (`avg`/`min`/`max`/`p5`/`p95`) and token totals are computed over successful rows only so failures cannot skew them.
 
-Full per-field semantics (including the per-invocation `failed_calls` rule and the `asyncio.gather` fan-out caveat on `call_duration`) live in the OpenAPI docs at `GET /docs`.
+Full per-field semantics (including how `failed_calls` is counted for multi-LLM-call invocations and the `asyncio.gather` fan-out caveat on `call_duration`) live in the OpenAPI docs at `GET /docs`.
 
 ## curl examples
