+++ pyproject.toml
@@ -177,6 +177,7 @@ ignore_imports = [
     "qfa.api.app -> qfa.adapters.presidio_anonymizer",
     "qfa.api.app -> qfa.adapters.tracking_llm",
     "qfa.api.app -> qfa.adapters.db",
+    "qfa.api.app -> qfa.adapters.usage_repository",
 ]
 
 [[tool.importlinter.contracts]]
