+++ src/qfa/api/app.py
@@ -17,13 +17,13 @@ from starlette.types import ASGIApp, Message, Receive, Scope, Send
 
 import qfa
 from qfa.adapters.db import (
-    SqlAlchemyUsageRepository,
     create_async_engine_from_settings,
     create_session_factory,
 )
 from qfa.adapters.llm_client import LiteLLMClient
 from qfa.adapters.presidio_anonymizer import PresidioAnonymizer
 from qfa.adapters.tracking_llm import TrackingLLMAdapter
+from qfa.adapters.usage_repository import SqlAlchemyUsageRepository
 from qfa.api.routes import router
 from qfa.api.routes_usage import router as usage_router
 from qfa.api.schemas import (
